Doggo
Split control out the a Game Controller object and stuff and things
