+++ Assets/prefabs/doggo.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 86e8f1c0786c6824cb2acdeb9d8946d5
+timeCreated: 1492170476
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
